backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "lightning>=2.5.1.post0",
    "pillow>=11.0.0",
    "sentence-transformers>=4.1.0",
    "torch>=2.8.0.dev20250520",
    "torchvision>=0.22.0.dev20250520",
    "tqdm>=4.66.5",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/nightly/cpu"
